hackathon_v1/ai_agent/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
openai==1.68.2
requests==2.32.3
anthropic==0.49.0
mcp==1.5.0
mcp-agent==0.0.10
